requirements.txt
requests
pyyaml
git+https://github.com/d3QUone/logger
